Added bower init Move electron_boilerplate library to directory lib
