+++ .gitignore
@@ -6,3 +6,4 @@ Thumbs.db
 /releases/
 /tmp/
 /.idea/
+/app/vendor
