+++ app/app.html
@@ -6,11 +6,11 @@
 
     <link href="./stylesheets/main.css" rel="stylesheet" type="text/css">
 
-    <script src="vendor/electron_boilerplate/env_config.js"></script>
-    <script src="vendor/electron_boilerplate/context_menu.js"></script>
-    <script src="vendor/electron_boilerplate/external_links.js"></script>
+    <script src="lib/electron_boilerplate/env_config.js"></script>
+    <script src="lib/electron_boilerplate/context_menu.js"></script>
+    <script src="lib/electron_boilerplate/external_links.js"></script>
 
-    <script src="vendor/require.js"></script>
+    <script src="vendor/requirejs/require.js"></script>
     <script>
         requirejs(['app'], function (app) {});
     </script>
