+++ app/hello_world/hello_universe.js
@@ -1,3 +1,5 @@
+"use strict";
+
 var greet = function () {
     return 'Hello Universe!';
 };
