+++ app/hello_world/hello_universe.spec.js
@@ -1,3 +1,5 @@
+"use strict";
+
 // Default imports test
 import greet from './hello_universe';
