+++ app/hello_world/hello_world.js
@@ -1,3 +1,5 @@
+"use strict";
+
 export var greet = function () {
     return 'Hello World!';
 };
