+++ app/hello_world/hello_world.spec.js
@@ -1,3 +1,5 @@
+"use strict";
+
 // Named imports test
 import { greet, bye } from './hello_world';
