+++ app/lib/electron_boilerplate/context_menu.js
