+++ app/lib/electron_boilerplate/dev_helper.js
