+++ app/lib/electron_boilerplate/external_links.js
