+++ app/lib/electron_boilerplate/window_state.js
