+++ app/main.js
@@ -2,9 +2,9 @@
 
 var app = require('app');
 var BrowserWindow = require('browser-window');
-var env = require('./vendor/electron_boilerplate/env_config');
-var devHelper = require('./vendor/electron_boilerplate/dev_helper');
-var windowStateKeeper = require('./vendor/electron_boilerplate/window_state');
+var env = require('./lib/electron_boilerplate/env_config');
+var devHelper = require('./lib/electron_boilerplate/dev_helper');
+var windowStateKeeper = require('./lib/electron_boilerplate/window_state');
 
 var mainWindow;
