+++ app/spec.html
@@ -4,18 +4,18 @@
     <meta charset="utf-8">
     <title>Jasmine Spec Runner</title>
 
-    <link rel="stylesheet" type="text/css" href="vendor/jasmine/jasmine.css">
+    <link rel="stylesheet" type="text/css" href="lib/jasmine/jasmine.css">
 
-    <script src="vendor/jasmine/jasmine.js"></script>
-    <script src="vendor/jasmine/jasmine-html.js"></script>
-    <script src="vendor/jasmine/boot.js"></script>
+    <script src="lib/jasmine/jasmine.js"></script>
+    <script src="lib/jasmine/jasmine-html.js"></script>
+    <script src="lib/jasmine/boot.js"></script>
     <script>
         // Overwrite defaul behaviour of boot.js. We can't start with onload event,
         // need to wait until requirejs loads all specs.
         var startJasmine = window.onload;
         window.onload = null;
     </script>
-    <script src="vendor/require.js"></script>
+    <script src="vendor/requirejs/require.js"></script>
 
     <script>
