+++ app/spec.js
@@ -2,8 +2,8 @@
 
 var app = require('app');
 var BrowserWindow = require('browser-window');
-var devHelper = require('./vendor/electron_boilerplate/dev_helper');
-var windowStateKeeper = require('./vendor/electron_boilerplate/window_state');
+var devHelper = require('./lib/electron_boilerplate/dev_helper');
+var windowStateKeeper = require('./lib/electron_boilerplate/window_state');
 
 var mainWindow;
